Added support for iterative keyword generation for larger datasets

+++ generate_keywords.py
@@ -1,3 +1,4 @@
+import os
 
 import pandas as pd
 import numpy as np
@@ -6,51 +7,37 @@ from dataset.yelp.loader import YelpDatasetLoader
 from models.nlp.KeyBERT import KeyBERTExtractor
 from models.nlp.yake import YakeExtractor
 
+output_dir = 'processed_df_cache'
+
+if not os.path.exists(output_dir):
+    os.makedirs(output_dir)
 
 amazon_dataloader = AmazonDatasetLoader()
 #yelp_dataloader = YelpDatasetLoader()
 
-df = amazon_dataloader.get_pandas_df()
+df_all = amazon_dataloader.get_pandas_df()
 
+filename = AmazonDatasetLoader.filenames[0].split('/')[-1].split('.')[0]
 
+df_pieces = np.array_split(df_all, 10)
 
+for i, df in enumerate(df_pieces):
+    current_part_path = f'{output_dir}/{filename}_{i}.gzip'
+    if os.path.exists(current_part_path):
+        print(f'-------------- Skipping {i} --------------')
+        continue
 
-#%%
+    print(f'-------------- Starting {i} --------------')
 
-df = KeyBERTExtractor().extract_keywords(df, {'top_n': 15, 'keyphrase_ngram_range': (1, 2)})
+    df = KeyBERTExtractor().extract_keywords(df, {'top_n': 15, 'keyphrase_ngram_range': (1, 1)})
 
-#%%
+    df = YakeExtractor().extract_keywords(df)
+    df.to_pickle(current_part_path)
 
-df.to_pickle('Digital_Music_5_with_extracted_topics.gzip')
-df = YakeExtractor().extract_keywords(df)
-
-
-
-#%%
-
-#reviews = df['review'].to_list()
-#res = KeyBERTExtractor().model.extract_keywords(reviews)
-
-
-#%%
-
-from models.nlp.tfidf import tfidfExtractor
-
-
-#df = amazon_dataloader.get_processed_pandas_df()
-
-df = tfidfExtractor().extract_keywords(df)
-
-#%%
-
-df.to_pickle('Digital_Music_5_with_extracted_topics.gzip')
-
-#%%
-
-
-
-#%%
-
-
+    """
+    from models.nlp.tfidf import tfidfExtractor
+    df = tfidfExtractor().extract_keywords(df)
+    df.to_pickle('Digital_Music_5_with_extracted_topics.gzip')
+    """
